+++ packages/ciphers/src/node-crypto.d.ts
@@ -5,6 +5,32 @@ declare module 'node:crypto' {
     final(): Uint8Array;
   }
 
+  interface CipherGCM extends Cipher {
+    setAAD(buffer: Uint8Array): this;
+    getAuthTag(): Uint8Array;
+  }
+
+  interface DecipherGCM extends Cipher {
+    setAAD(buffer: Uint8Array): this;
+    setAuthTag(buffer: Uint8Array): this;
+  }
+
+  interface CipherGCMTypes {
+    authTagLength: number;
+  }
+
   export function createCipheriv(algorithm: string, key: Uint8Array, iv: Uint8Array | null): Cipher;
+  export function createCipheriv(
+    algorithm: string,
+    key: Uint8Array,
+    iv: Uint8Array,
+    options: CipherGCMTypes,
+  ): CipherGCM;
   export function createDecipheriv(algorithm: string, key: Uint8Array, iv: Uint8Array | null): Cipher;
+  export function createDecipheriv(
+    algorithm: string,
+    key: Uint8Array,
+    iv: Uint8Array,
+    options: CipherGCMTypes,
+  ): DecipherGCM;
 }